requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
blas=1.0=mkl
brotli=1.0.9=h8ffe710_7
brotli-bin=1.0.9=h8ffe710_7
brotlipy=0.7.0=py38h294d835_1004
bzip2=1.0.8=h8ffe710_4
ca-certificates=2021.10.8=h5b45459_0
certifi=2021.10.8=py38haa244fe_2
cffi=1.15.0=py38hd8c33c5_0
charset-normalizer=2.0.12=pyhd8ed1ab_0
colorama=0.4.4=pyh9f0ad1d_0
cryptography=36.0.0=py38h21b164f_0
cudatoolkit=11.3.1=h59b6b97_2
cycler=0.11.0=pyhd8ed1ab_0
fonttools=4.33.3=py38h294d835_0
freetype=2.10.4=h546665d_1
icu=69.1=h0e60522_0
idna=3.3=pyhd8ed1ab_0
intel-openmp=2022.0.0=h57928b3_3663
jbig=2.1=h8d14728_2003
joblib=1.1.0=pyhd8ed1ab_0
jpeg=9e=h8ffe710_1
kiwisolver=1.4.2=py38hbd9d945_1
lcms2=2.12=h2a16943_0
lerc=3.0=h0e60522_0
libblas=3.9.0=14_win64_mkl
libbrotlicommon=1.0.9=h8ffe710_7
libbrotlidec=1.0.9=h8ffe710_7
libbrotlienc=1.0.9=h8ffe710_7
libcblas=3.9.0=14_win64_mkl
libclang=13.0.1=default_h81446c8_0
libdeflate=1.10=h8ffe710_0
libffi=3.4.2=h8ffe710_5
liblapack=3.9.0=14_win64_mkl
libpng=1.6.37=h1d00b33_2
libtiff=4.3.0=hc4061b1_3
libuv=1.43.0=h8ffe710_0
libwebp=1.2.2=h57928b3_0
libwebp-base=1.2.2=h8ffe710_1
libxcb=1.13=hcd874cb_1004
libzlib=1.2.11=h8ffe710_1014
lz4-c=1.9.3=h8ffe710_1
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
matplotlib=3.5.2=py38haa244fe_0
matplotlib-base=3.5.2=py38he529843_0
mkl=2022.0.0=h0e2418a_796
msys2-conda-epoch=20160418=1
munkres=1.1.4=pyh9f0ad1d_0
numpy=1.22.3=py38h1d2777f_2
openjpeg=2.4.0=hb211442_1
openssl=1.1.1o=h8ffe710_0
packaging=21.3=pyhd8ed1ab_0
pillow=9.1.0=py38hd8e0db4_2
pip=22.0.4=pyhd8ed1ab_0
pthread-stubs=0.4=hcd874cb_1001
pycparser=2.21=pyhd8ed1ab_0
pyopenssl=22.0.0=pyhd8ed1ab_0
pyparsing=3.0.8=pyhd8ed1ab_0
pyqt=5.12.3=py38haa244fe_8
pyqt-impl=5.12.3=py38h885f38d_8
pyqt5-sip=4.19.18=py38h885f38d_8
pyqtchart=5.12=py38h885f38d_8
pyqtwebengine=5.12.1=py38h885f38d_8
pysocks=1.7.1=py38haa244fe_5
python=3.8.13=h9a09f29_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.8=2_cp38
pytorch=1.11.0=py3.8_cuda11.3_cudnn8_0
pytorch-mutex=1.0=cuda
qt=5.12.9=h556501e_6
requests=2.27.1=pyhd8ed1ab_0
scikit-learn=1.0.2=py38hb60ee80_0
scipy=1.8.0=py38ha1292f7_1
setuptools=62.1.0=py38haa244fe_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.38.3=h8ffe710_0
tbb=2021.5.0=h2d74725_1
threadpoolctl=3.1.0=pyh8a188c0_0
tk=8.6.12=h8ffe710_0
torchaudio=0.11.0=py38_cu113
torchvision=0.12.0=py38_cu113
tornado=6.1=py38h294d835_3
tqdm=4.64.0=pyhd8ed1ab_0
typing_extensions=4.2.0=pyha770c72_1
ucrt=10.0.20348.0=h57928b3_0
unicodedata2=14.0.0=py38h294d835_1
urllib3=1.26.9=pyhd8ed1ab_0
vc=14.2=hb210afc_6
vs2015_runtime=14.29.30037=h902a5da_6
wheel=0.37.1=pyhd8ed1ab_0
win_inet_pton=1.1.0=py38haa244fe_4
xorg-libxau=1.0.9=hcd874cb_0
xorg-libxdmcp=1.1.3=hcd874cb_0
xz=5.2.5=h62dcd97_1
zlib=1.2.11=h8ffe710_1014
zstd=1.5.2=h6255e5f_0
